add mermaid graph generation
