+++ logos-codegen/Cargo.toml
@@ -21,8 +21,6 @@ debug = []
 fuzzing = []
 # Don't use or generate unsafe code
 forbid_unsafe = []
-# Enables graphviz DOT file generation
-graphviz = []
 
 [lib]
 bench = false
