+++ logos-codegen/src/graph/mod.rs
@@ -7,8 +7,8 @@ use std::ops::Index;
 
 use fnv::FnvHasher;
 
-#[cfg(feature = "graphviz")]
-mod dot;
+#[cfg(feature = "debug")]
+mod export;
 mod fork;
 mod impls;
 mod meta;
